add restriction for deleting and editing tasks

+++ todolist_app/views.py
@@ -33,7 +33,11 @@ def todolist(request):
 @login_required
 def delete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.delete()
+
+    if task.owner == request.user:
+        task.delete()
+    else:
+        messages.error(request, ("Access denied, you are not allowed!"))
 
     return redirect('todolist')
 
@@ -41,8 +45,11 @@ def delete_task(request, task_id):
 @login_required
 def complete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = True
-    task.save()
+    if task.owner == request.user:
+        task.done = True
+        task.save()
+    else:
+        messages.error(request, ("Access denied, you are not allowed!"))
 
     return redirect('todolist')
 
@@ -50,8 +57,11 @@ def complete_task(request, task_id):
 @login_required
 def pending_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = False
-    task.save()
+    if task.owner == request.user:
+        task.done = False
+        task.save()
+    else:
+        messages.error(request, ("Access denied, you are not allowed!"))
 
     return redirect('todolist')
